requirements.txt
llama-index==0.7.11
streamlit==1.25.0
pypdf==3.12.2
PyPDF2==3.0.1
python-dotenv==1.0.0
